Cleanup of sources and resources file. Also added the example secrets.
