+++ pipelines/.dlt/example.secrets.toml
@@ -27,6 +27,12 @@ secret_str="secret string"  # a string secret
 secret_key="key string"
 key_index=1
 
+# google analytics, the client email is the email of the service account.
+[sources.google_analytics.google_analytics.credentials]
+project_id = ""
+client_email = ""
+private_key = ""
+
 ## google sheets pipeline
 [sources.google_sheets.credentials]
 project_id=""
